penamabahan fitur dan perbaikan layout

+++ app/Database/Migrations/2026-06-16-000001_AddFotoToUserTable.php
@@ -8,23 +8,27 @@ class AddFotoToUserTable extends Migration
 {
     public function up()
     {
-        if (! $this->db->fieldExists('foto', 'user')) {
-            $fields = [
-                'foto' => [
-                    'type'       => 'VARCHAR',
-                    'constraint' => '255',
-                    'null'       => true,
-                    'default'    => null,
-                ],
-            ];
-            $this->forge->addColumn('user', $fields);
+        if ($this->db->tableExists('user')) {
+            if (! $this->db->fieldExists('foto', 'user')) {
+                $fields = [
+                    'foto' => [
+                        'type'       => 'VARCHAR',
+                        'constraint' => '255',
+                        'null'       => true,
+                        'default'    => null,
+                    ],
+                ];
+                $this->forge->addColumn('user', $fields);
+            }
         }
     }
 
     public function down()
     {
-        if ($this->db->fieldExists('foto', 'user')) {
-            $this->forge->dropColumn('user', 'foto');
+        if ($this->db->tableExists('user')) {
+            if ($this->db->fieldExists('foto', 'user')) {
+                $this->forge->dropColumn('user', 'foto');
+            }
         }
     }
 }
