+++ app/Database/Migrations/2026-06-16-000002_AddAkreditasiToSekolahTable.php
@@ -8,22 +8,21 @@ class AddAkreditasiToSekolahTable extends Migration
 {
     public function up()
     {
-        if (! $this->db->fieldExists('akreditasi', 'sekolah')) {
-            $fields = [
+        if ($this->db->tableExists('sekolah') && !$this->db->fieldExists('akreditasi', 'sekolah')) {
+            $this->forge->addColumn('sekolah', [
                 'akreditasi' => [
                     'type'       => 'ENUM',
                     'constraint' => ['A', 'B', 'C', 'Belum Terakreditasi', 'Tidak Diketahui'],
                     'null'       => false,
                     'default'    => 'Belum Terakreditasi',
                 ],
-            ];
-            $this->forge->addColumn('sekolah', $fields);
+            ]);
         }
     }
 
     public function down()
     {
-        if ($this->db->fieldExists('akreditasi', 'sekolah')) {
+        if ($this->db->tableExists('sekolah') && $this->db->fieldExists('akreditasi', 'sekolah')) {
             $this->forge->dropColumn('sekolah', 'akreditasi');
         }
     }
